Use number of nodes to determine if binary tree is perfect.
Initially I had used the height to determine if a binary tree
was a perfect binary tree. It was bugy and so I tried using
the number of nodes in the left and right sub-trees of the root
node. A perfect binary tree has equal number of nodes in each
sub-tree.