Change logic of finding height of a binary tree
One of the tests on the alx website failed yet my
output was correct, so I changed the logic to try
and pass the test. The test was closed and hence
I had to guess at what the issue was. I used a different
approach of making the height of an empty tree equal to -1
and height of a leaf node equal to zero.